requirements.txt
django>=4.0.0,<4.1.0
django-environ
django-cors-headers
djangorestframework
djangorestframework-simplejwt

black
pylint
pylint-django
